stok çıkış iş emri gönder butonu altı düzenlendi.(web service)

+++ Chep.Core/VwWsWorkOrderInvoiceDetails.cs
@@ -30,7 +30,7 @@ namespace Chep.Core
         public string PkgType { get; set; }
         public string CommclDesc { get; set; }
         public int NumberOfPackages { get; set; }
-        public long? ItemNumber { get; set; }
+        public int? ItemNumber { get; set; }
         public string ProducerCompanyNo { get; set; }
         public string ProducerCompany { get; set; }
         public string IncentiveLineNo { get; set; }
